Refactor LocateByLayer in a module
